% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{align}
\alias{align}
\title{Align two binary ordered trees}
\usage{
align(T1, T2, cost_matrix)
}
\arguments{
\item{T1}{A binary tree}

\item{T2}{A binary tree}

\item{cost_matrix}{Cost matrix between each pair of nodes in T1 and T2}
}
\value{
An alignment object
}
\description{
Align two binary ordered trees
}
