% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{as_binary_tree.default}
\alias{as_binary_tree.default}
\title{Convert a newick string as a binary tree object}
\usage{
\method{as_binary_tree}{default}(newick_string)
}
\arguments{
\item{newick_string}{A newick string}
}
\value{
A binary tree object
}
\description{
Convert a newick string as a binary tree object
}
