% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{as_binary_tree.hclust}
\alias{as_binary_tree.hclust}
\title{Convert a hclust object as a binary tree object}
\usage{
\method{as_binary_tree}{hclust}(hclust_obj, data, membership,
  method = "median")
}
\arguments{
\item{hclust_obj}{A hclust object}

\item{data}{A data matrix}

\item{membership}{A membership matrix for leaves}

\item{method}{A string indicating method for summarizing the nodes}
}
\value{
A list containing the binary tree object and a vector of summary statistics on each node
}
\description{
Convert a hclust object as a binary tree object
}
