% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{as_binary_tree.phylo}
\alias{as_binary_tree.phylo}
\title{Convert a phylo object as a binary tree object}
\usage{
\method{as_binary_tree}{phylo}(phylo_obj)
}
\arguments{
\item{phylo_obj}{A phylo object}
}
\value{
A binary tree object
}
\description{
Convert a phylo object as a binary tree object
}
