% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{as_binary_tree}
\alias{as_binary_tree}
\title{Generic method for convert trees to binary tree}
\usage{
as_binary_tree(x, ...)
}
\arguments{
\item{...}{Additional arguments for conversion}

\item{object}{An object to be converted to a binary_tree}
}
\description{
`as_binary_tree` is a generic function for converting various objects into a 
binary_tree object
}
