% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{as_phylo}
\alias{as_phylo}
\title{Convert a binary tree as phylo object}
\usage{
as_phylo(binary_tree)
}
\arguments{
\item{binary_tree}{A binary tree object}
}
\value{
A phylo object
}
\description{
Convert a binary tree as phylo object
}
