% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{binary_tree}
\alias{binary_tree}
\title{Create a binary tree object}
\usage{
binary_tree(label, left = NULL, right = NULL)
}
\arguments{
\item{label}{The label of the node}

\item{left}{The left binary tree}

\item{right}{The right binary tree}
}
\value{
A list representing the tree
}
\description{
Create a binary tree object
}
