% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{binary_tree_node}
\alias{binary_tree_node}
\title{create a binary tree node}
\usage{
binary_tree_node(label, left = NULL, right = NULL)
}
\arguments{
\item{label}{The label of the node}

\item{left}{The left tree node}

\item{right}{The right tree node}
}
\value{
A list representing the tree
}
\description{
create a binary tree node
}
