% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_functions.R
\name{clustertree_construction}
\alias{clustertree_construction}
\title{Construct clustering tree from proportion graph}
\usage{
clustertree_construction(clusters.aggregate)
}
\arguments{
\item{clusters.aggregate}{A data matrix with cluster membership at different levels (column names should be K1, K2, ...)}
}
\value{
A list containing a proportion graph and a trimmed version keeping only the incident edge with highest weight for each node
}
\description{
Construct clustering tree from proportion graph
}
