% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{concatLabels}
\alias{concatLabels}
\title{A helper function that takes node id and return the name of all leaves that
are its children}
\usage{
concatLabels(unique_labels)
}
\arguments{
\item{unique_labels}{a $c \times 2$ data frame with column names id (cluster id) and
desc (the name of any leave that is its child)}
}
\value{
a data frame that contains node id and concatenated names of all its children
leaves
}
\description{
A helper function that takes node id and return the name of all leaves that
are its children
}
