% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph_clustering.R
\name{construct_hierarchy}
\alias{construct_hierarchy}
\title{Construct Hierarchy based on Cluster Mean}
\usage{
construct_hierarchy(data, membership, dist_method = "euclidean",
  hclust_method = "complete", foldchange = T)
}
\arguments{
\item{data}{A data matrix to compute cluster mean..}

\item{membership}{A vector of cluster membership.}
}
\value{
A hclust object.
}
\description{
Construct cluster hierarchy based on the mean of each cluster.
}
