% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph_clustering.R
\name{ConstructHierarchy}
\alias{ConstructHierarchy}
\title{Construct Hierarchy based on Cluster Mean}
\usage{
ConstructHierarchy(data, membership, func)
}
\arguments{
\item{data}{A data matrix to compute cluster mean..}

\item{membership}{A vector of cluster membership.}
}
\value{
A hclust object.
}
\description{
Construct cluster hierarchy based on the mean of each cluster.
}
