% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_matrix_functions.R
\name{cost_matrix}
\alias{cost_matrix}
\title{Create cost matrix based on distance}
\usage{
cost_matrix(T1_stats, T2_stats, lambda = NA, method = "spearman")
}
\arguments{
\item{T1_stats}{A matrix with each row as features for each node}

\item{lambda}{The cost of aligning a node with a empty node}

\item{method}{The method used to compute correlation (spearman, pearson or euclidean)}

\item{T2}{A matrix with each row as features for each node}
}
\value{
A pairwise distance matrix
}
\description{
Create cost matrix based on distance
}
