% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{create_align_object}
\alias{create_align_object}
\title{Create alignment object}
\usage{
create_align_object(T1, T2, cost_matrix)
}
\arguments{
\item{T1}{tree 1}

\item{T2}{tree 2}

\item{cost_matrix}{The cost matrix}
}
\value{
An alignment object
}
\description{
Create alignment object
}
