% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{DE_level_k}
\alias{DE_level_k}
\title{sensitive DE at level k of the hierarchy}
\usage{
DE_level_k(cluster_name, expression, q_thresh, FC_thresh)
}
\arguments{
\item{cluster_name}{the name of the clusters that each cell belongs to}

\item{expression}{$m \times n$ expression matrix for $m$ genes and $n$ cells}

\item{q_thresh}{q value threshold for DE analysis}

\item{FC_thresh}{fold change threshold for DE analysis}
}
\value{
differentially expressed genes for each cluster
}
\description{
sensitive DE at level k of the hierarchy
}
