% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_matrix_functions.R
\name{euc_cost_matrix}
\alias{euc_cost_matrix}
\title{Create cost matrix with 1 - cor}
\usage{
euc_cost_matrix(T1_stats, T2_stats, lambda = NA)
}
\arguments{
\item{T1_stats}{A matrix with each row as features for each node}

\item{lambda}{The cost of aligning a node with a empty node}

\item{T2}{A matrix with each row as features for each node}
}
\value{
A pairwise distance matrix (1 - cor)
}
\description{
Create cost matrix with 1 - cor
}
