% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_functions.R
\name{fill_matrix}
\alias{fill_matrix}
\title{Fill T matrix and F map}
\usage{
fill_matrix(align_obj)
}
\arguments{
\item{align_obj}{An alignment object}
}
\value{
An alignment object with T matrix and F map filled
}
\description{
Fill T matrix and F map
}
