% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{get_children}
\alias{get_children}
\title{get the children labels}
\usage{
get_children(node)
}
\arguments{
\item{node}{A node object}
}
\value{
A character vector of children's names
}
\description{
get the children labels
}
