% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{get_HVG}
\alias{get_HVG}
\title{Get highly variable genes}
\usage{
get_HVG(filtered.normalized, dir_path, min.mean, min.frac.cells,
  min.dispersion.scaled, remove_RP = TRUE)
}
\arguments{
\item{filtered.normalized}{Filtered and normalized data matrix}

\item{dir_path}{Path to the directory where the figure is output}

\item{min.mean}{The minimum mean for filtering criterion}

\item{min.frac.cells}{The fraction of cells that are not zero for genes}

\item{min.dispersion.scaled}{The minimum scaled dispersion for the genes}
}
\value{
The names of the genes that are defined as highly variable based on the criterion
}
\description{
Get highly variable genes
}
