% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_jaccard}
\alias{get_jaccard}
\title{return the pairwise Jaccard coefficients between two experiemnts}
\usage{
get_jaccard(inputA, inputB)
}
\arguments{
\item{inputA}{a data frame with column names node (the name of the cluster) and
gene (all overexpressed gene of the node)}

\item{inputB}{same format as inputA}
}
\value{
a list containing the Jaccard coefficients in long and wide form
}
\description{
return the pairwise Jaccard coefficients between two experiemnts
}
