% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{initialize}
\alias{initialize}
\title{Initialize the matrix to store alignment score}
\usage{
initialize(T1, T2, cost_matrix)
}
\arguments{
\item{T1}{tree 1}

\item{T2}{tree 2}

\item{cost_matrix}{The cost matrix}
}
\value{
A list of T matrix and F map
}
\description{
Initialize the matrix to store alignment score
}
