% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cut_tree.R
\name{makeCutTree}
\alias{makeCutTree}
\title{return the output of cutree on an hclust object generated from the first
ten principal components of the data slot of a seurat object.}
\usage{
makeCutTree(seurat, n = 16)
}
\arguments{
\item{seurat}{a seurat-class objec}

\item{n}{an integer (default 16) of the maximum number of splits shown}
}
\value{
a matrix with group memberships from 1 to n
}
\description{
return the output of cutree on an hclust object generated from the first
ten principal components of the data slot of a seurat object.
}
