% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph_clustering.R
\name{PCPermutationTest}
\alias{PCPermutationTest}
\title{Test for significant PCs by Permutation Parallel Analysis}
\usage{
PCPermutationTest(dat, B = 100, threshold = 0.05, randomized = F,
  verbose = TRUE, seed = 1, max.pc = 100, n.cores = 1,
  center = T, scale = T)
}
\arguments{
\item{dat}{A data matrix containing the raw data (each column should be a feature).}

\item{B}{The number of permutations.}

\item{threshold}{p-value for significance.}

\item{seed}{An optional seed.}
}
\value{
A list containing the number of signficant PCs and p value
}
\description{
Estimate a number of significant principal components from a permutation test
}
