% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph_clustering.R
\name{PhenoGraphVarRes}
\alias{PhenoGraphVarRes}
\title{PhenoGraph Clustering with Various Resolution}
\usage{
PhenoGraphVarRes(data, K)
}
\arguments{
\item{data}{A input data matrix for Rphenograph.}

\item{K}{A numeric vector of resolutions.}
}
\value{
A list containing a data matrix of membership at each K and a

A list containing a data matrix of membership at each K and a
vector of clustering modularity
}
\description{
Run RPhenograph with different resolution and save the membership as a data
frame with column names as the k parameter.
}
\examples{
iris_unique = unique(iris) # Remove duplicates
data = as.matrix(iris_unique[,1:4])
K = seq(10, 50, length.out=5)
res = PhenoGraphVarRes(data, K=K)
res
}
