% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph_clustering.R
\name{plot.hierarchy}
\alias{plot.hierarchy}
\title{Plot Hierarchy Constructed}
\usage{
\method{plot}{hierarchy}(x, ha_row = NULL, show_row_names = T,
  show_col_names = F)
}
\arguments{
\item{x}{A hierarchy object.}
}
\value{
A heatmap.2 object
}
\description{
Plot Hierarchy Constructed
}
