% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{postorder}
\alias{postorder}
\title{Get labels by postorder traversal}
\usage{
postorder(x)
}
\arguments{
\item{x}{A root node of tree}
}
\value{
labels of tree in postorder
}
\description{
Get labels by postorder traversal
}
