% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{postorder_labels}
\alias{postorder_labels}
\title{Get labels by postorder traversal}
\usage{
postorder_labels(x)
}
\arguments{
\item{x}{A binary tree}
}
\value{
labels of tree in postorder
}
\description{
Get labels by postorder traversal
}
