% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess 10X dataset}
\usage{
preprocess_data(path, min.cells = 3, min.genes = 200,
  nGene_thresh = NULL, percent_mito_thresh = 0.98,
  mito_thresh = NULL, x.low.cutoff = 0.1, x.high.cutoff = 8,
  y.cutoff = 1, regress = F)
}
\arguments{
\item{path}{The path to the dataset}

\item{min.cells}{The minimum number of cells for each gene}

\item{min.genes}{The minimum number of genes for each cell}

\item{nGene_thresh}{The quantile threshold for multilplets based on nGene}

\item{percent_mito_thresh}{The quantile threshold for multilplets based on mitochondrial gene expression percentage}
}
\value{
A Seurat object with data normalized, scaled and PCA performed
}
\description{
Preprocess 10X dataset
}
\author{
Meng Wang
}
