% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{qc}
\alias{qc}
\title{Preprocess two 10X Genomics pan T cell datasets and one fibroblast
reprogamming dataset. Peform QC and normalize.}
\usage{
qc(input_data)
}
\arguments{
\item{input_data}{the input data.frame with genes in row.names and
cell names as columns}
}
\value{
a data.frame with QC filtered cells, genes, and normalized counts
To run on the three test datasets, load raw data:
library(dyno)
library(SingleCellExperiment)
library(scater)
options(stringsAsFactors = FALSE)
library(scran)
library(mvoutlier)
library(Seurat)
data("fibroblast_reprogramming_treutlein")
fibro <- as.data.frame(t(fibroblast_reprogramming_treutlein$counts))
load('t3.Seurat.RData')
load('t4.Seurat.RData')
t3.data <- as.data.frame(as.matrix(t3@raw.data))
t4.data <- as.data.frame(as.matrix(t4@raw.data))
}
\description{
Preprocess two 10X Genomics pan T cell datasets and one fibroblast
reprogamming dataset. Peform QC and normalize.
}
