% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{reset_pca}
\alias{reset_pca}
\title{Perform PCA on normalized counts.}
\usage{
reset_pca(input_data)
}
\arguments{
\item{input_data}{the normalized counts to run PCA on.}
}
\value{
PCA components.
}
\description{
Perform PCA on normalized counts.
}
