% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{sens_DE}
\alias{sens_DE}
\title{sensitive DE at all levels of the hierarchy}
\usage{
sens_DE(expression.all, cell_ancestor_matrix, q_thresh, FC_thresh)
}
\arguments{
\item{expression.all}{the $m$ genes by $n$ cells expression matrix}

\item{cell_ancestor_matrix}{hierarchy of the cells}

\item{q_thresh}{q value threshold for DE analysis}

\item{FC_thresh}{fold change threshold for DE analysis}
}
\value{
differentially expressed genes for each node in the hierarchy
}
\description{
sensitive DE at all levels of the hierarchy
}
