% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{sensDE_limma}
\alias{sensDE_limma}
\title{Use limma to find differentially expressed genes}
\usage{
sensDE_limma(expression, unique_cluster_membership, q_thresh = 0.1,
  FC_thresh = 1.2)
}
\arguments{
\item{expression}{$m \times n$ expression matrix with $m$ genes and $n$ cells}

\item{unique_cluster_membership}{a logical vector of size $n$ that indicates membership}

\item{q_thresh}{q value threshold for DE analysis}

\item{FC_thresh}{fold change threshold for DE analysis}
}
\description{
Use limma to find differentially expressed genes
}
