% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_signature.R
\name{SimpleDE}
\alias{SimpleDE}
\title{Differential expression that iteratively compares one cluster against all the others}
\usage{
SimpleDE(bottom_id, normalized, q_thresh = 0.1, FC_thresh = 1.2)
}
\arguments{
\item{bottom_id}{the cluster id of each cell}

\item{normalized}{the normalized expression matrix with column ordered as the bottom_id}

\item{q_thresh}{q value threshold for DE analysis}

\item{FC_thresh}{fold change threshold for DE analysis}
}
\value{
the list of genes that are overexpressed for each cluster against all others
}
\description{
Differential expression that iteratively compares one cluster against all the others
}
