% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_alignment_configuration}
\alias{test_alignment_configuration}
\title{Test alignment method}
\usage{
test_alignment_configuration(T3K_tree, T4K_tree, configuration)
}
\arguments{
\item{configuration}{A list of configuration for hierarchy construction}

\item{T3K}{a hierarchy object (for example T3K)}

\item{T4K}{A hierarchy object (for example T4K)}
}
\description{
Test alignment method
}
