% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_hierarchy_construction}
\alias{test_hierarchy_construction}
\title{Test hierarchy construction}
\usage{
test_hierarchy_construction(T3K, T4K, configuration, foldchange = T)
}
\arguments{
\item{T3K}{a hierarchy object (for example T3K)}

\item{T4K}{A hierarchy object (for example T4K)}

\item{configuration}{A list of configuration for hierarchy construction}
}
\description{
Test hierarchy construction
}
