% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_helper.R
\name{word_ancestor_matrix}
\alias{word_ancestor_matrix}
\title{Given a ancestor matrix of the hierarchy and corresponding description
of the leaves, return a name for each internal nodes}
\usage{
word_ancestor_matrix(id_matrix, cell_desc)
}
\arguments{
\item{id_matrix}{a m by k matrix that indicates the membership of
nodes at each level ($m$ is the number of cells and $k$ is the number of 
splits)}

\item{cell_desc}{a vector of size m that indicates the annotation at the
bottom of the hierarchy}
}
\value{
a ancestor matrix replacing id with corresponding name according to
last number
}
\description{
Given a ancestor matrix of the hierarchy and corresponding description
of the leaves, return a name for each internal nodes
}
