% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_functions.R
\name{initialize}
\alias{initialize}
\title{Intialize the T matrix and F map}
\usage{
initialize(align_object)
}
\arguments{
\item{align_object}{An alignment object}
}
\value{
An alignment object with T matrix and F map initialized
}
\description{
Intialize the T matrix and F map
}
